04-Image-Processing-pipeline/lambdas/resize/requirements.txt
logger
pillow